requirements.txt
dotenv
networkx
shap
google-genai
